First Commit
Data: Includes the data that was used for the visualizations
Notebooks: Includes the Jupyter Notebooks with the code. Please notice that only Matplotlib was used, without other libraries such as Seaborn or Plotly. This is due to project restrictions.
Images: Contains the finished results in .png format.
Docs: Contains short ReadMe and full Technical Report.